Move test over to the appropriate place
